requirements.txt
caldav >= 0.11.0
pytest-mock
python-dotenv
tuya-iot-py-sdk
